refactor of the server and first two tests
